fix(build): make local & CI build green on nightly 1.100
- Cargo.toml x86_64 0.15 pic8259 0.11 pc-keyboard 0.8 (fix Step trait)
- .cargo/config: json-target-spec=true + build-std
- gdt.rs: add_entry -> append + Segment trait (0.15 API)
- interrupts.rs: idt[u8] not usize (0.15)
- keyboard.rs: spin::Once not conquer-once (no_std)
- scheduler.rs/morphic.rs: alloc imports
- deck.rs/wasm.rs/synapse.rs: ToString/Clone fixes
- main.rs: remove legacy mods + BootInfo field fix
- x86_64-nimbus_os.json: int types + features empty
- remove package.metadata.bootloader target (use .cargo/config)
- built nimbus_os.bin 1.25MB, http.server 8000 running (safe, no OS change)

+++ src/interrupts.rs
@@ -24,8 +24,8 @@ lazy_static! {
         unsafe {
             idt.double_fault.set_handler_fn(double_fault_handler);
         }
-        idt[InterruptIndex::Timer as usize].set_handler_fn(timer_handler);
-        idt[InterruptIndex::Keyboard as usize].set_handler_fn(keyboard_handler);
+        idt[InterruptIndex::Timer as u8].set_handler_fn(timer_handler);
+        idt[InterruptIndex::Keyboard as u8].set_handler_fn(keyboard_handler);
         idt.page_fault.set_handler_fn(page_fault_handler);
         idt.general_protection_fault.set_handler_fn(gpf_handler);
         idt
